feel lazy to commit

+++ springboot-jwt/src/main/java/com/green/springbootjwt/repo/ItemRepository.java
@@ -3,5 +3,5 @@ package com.green.springbootjwt.repo;
 import com.green.springbootjwt.entity.Item;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ItemRepository extends CrudRepository<Item,Integer> {
+public interface ItemRepository extends CrudRepository<Item,Long> {
 }
